:bug: Consolidate theme logic into unified reactive ThemeState
