+++ app/src/commonMain/kotlin/com/crosspaste/ui/theme/ThemeDetector.kt
@@ -26,6 +26,52 @@ data class ThemeState(
     val colorScheme: ColorScheme,
 ) {
 
+    companion object {
+        fun createThemeState(
+            themeColor: ThemeColor,
+            colorContrast: ColorContrast,
+            isFollowSystem: Boolean,
+            isUserInDark: Boolean,
+            isSystemInDark: Boolean,
+        ): ThemeState {
+            val isCurrentThemeDark = if (isFollowSystem) isSystemInDark else isUserInDark
+            val currentColorScheme =
+                if (isCurrentThemeDark) {
+                    getDarkColorSchema(themeColor, colorContrast)
+                } else {
+                    getLightColorSchema(themeColor, colorContrast)
+                }
+            return ThemeState(
+                themeColor = themeColor,
+                colorContrast = colorContrast,
+                isFollowSystem = isFollowSystem,
+                isUserInDark = isUserInDark,
+                isSystemInDark = isSystemInDark,
+                colorScheme = currentColorScheme,
+            )
+        }
+
+        private fun getLightColorSchema(
+            themeColor: ThemeColor,
+            contrast: ColorContrast,
+        ): ColorScheme =
+            when (contrast) {
+                ColorContrast.Standard -> themeColor.lightColorScheme
+                ColorContrast.Medium -> themeColor.lightMediumContrastColorScheme
+                ColorContrast.High -> themeColor.lightHighContrastColorScheme
+            }
+
+        private fun getDarkColorSchema(
+            themeColor: ThemeColor,
+            contrast: ColorContrast,
+        ): ColorScheme =
+            when (contrast) {
+                ColorContrast.Standard -> themeColor.darkColorScheme
+                ColorContrast.Medium -> themeColor.darkMediumContrastColorScheme
+                ColorContrast.High -> themeColor.darkHighContrastColorScheme
+            }
+    }
+
     val isCurrentThemeDark: Boolean
         get() = if (isFollowSystem) isSystemInDark else isUserInDark
 }
